+++ sdwan_features_system.tf
@@ -38,7 +38,8 @@ resource "sdwan_system_aaa_feature" "system_aaa_feature" {
       address             = server.address
       auth_port           = try(server.authentication_port, null)
       auth_port_variable  = try("{{${server.authentication_port_variable}}}", null)
-      key                 = server.key
+      key                 = try(server.key, null)
+      key_variable        = try("{{${server.key_variable}}}", null)
       key_type            = try(server.key_type, null)
       key_type_variable   = try("{{${server.key_type_variable}}}", null)
       retransmit          = try(server.retransmit, null)
@@ -55,7 +56,8 @@ resource "sdwan_system_aaa_feature" "system_aaa_feature" {
     group_name = try("tacacs-${group.vpn}", "tacacs-0-0")
     servers = !can(group.servers) ? null : [for server in group.servers : {
       address          = server.address
-      key              = server.key
+      key              = try(server.key, null)
+      key_variable     = try("{{${server.key_variable}}}", null)
       port             = try(server.port, null)
       port_variable    = try("{{${server.port_variable}}}", null)
       timeout          = try(server.timeout, null)
@@ -73,8 +75,9 @@ resource "sdwan_system_aaa_feature" "system_aaa_feature" {
     privilege          = try(user.privilege, null)
     privilege_variable = try("{{${user.privilege_variable}}}", null)
     public_keys = try(length(user.public_key_chains) == 0, true) ? null : [for public_key in user.public_key_chains : {
-      key_type   = "ssh-rsa"
-      key_string = public_key
+      key_type            = "ssh-rsa"
+      key_string          = try(public_key.key, null)
+      key_string_variable = try("{{${public_key.key_variable}}}", null)
     }]
   }]
 }
@@ -405,7 +408,6 @@ resource "sdwan_system_mrf_feature" "system_mrf_feature" {
   feature_profile_id           = sdwan_system_feature_profile.system_feature_profile[each.value.name].id
   enable_migration_to_mrf      = try(each.value.mrf.migration_to_mrf, null)
   migration_bgp_community      = try(each.value.mrf.migration_bgp_community, null)
-  region_id                    = try(each.value.mrf.region_id, null)
   role                         = try(each.value.mrf.role, null)
   role_variable                = try("{{${each.value.mrf.role_variable}}}", null)
   secondary_region_id          = try(each.value.mrf.secondary_region_id, null)